review comments fix
