+++ kafka-streams-partitioners/avro-partitioners/src/test/avro/TestRecord.avdl
