+++ kafka-streams-partitioners/avro-partitioners/src/test/resources/log4j2.properties
@@ -1,0 +1,10 @@
+status = error
+name = PropertiesConfig
+
+appender.console.type = Console
+appender.console.name = STDOUT
+appender.console.layout.type = PatternLayout
+appender.console.layout.pattern = %d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %c{1.} - %msg%n
+
+rootLogger.level = INFO
+rootLogger.appenderRef.stdout.ref = STDOUT
