--- kafka-streams-partitioners/src/main/java/org/hypertrace/core/kafkastreams/framework/partitioner/ConfigProvider.java
@@ -1,5 +1,0 @@
-package org.hypertrace.core.kafkastreams.framework.partitioner;
-
-interface ConfigProvider {
-  MultiLevelPartitionerConfig getConfig();
-}
