+++ kafka-streams-partitioners/weighted-group-partitioner/src/main/java/org/hypertrace/core/kafkastreams/framework/partitioner/PartitionerConfigServiceClient.java
@@ -1,0 +1,5 @@
+package org.hypertrace.core.kafkastreams.framework.partitioner;
+
+interface PartitionerConfigServiceClient {
+  WeightedGroupProfile getConfig(String profileName);
+}
