+++ kafka-streams-partitioners/weighted-group-partitioner/src/main/java/org/hypertrace/core/kafkastreams/framework/partitioner/RoundRobinPartitioner.java
@@ -1,0 +1,13 @@
+package org.hypertrace.core.kafkastreams.framework.partitioner;
+
+import java.util.concurrent.atomic.AtomicInteger;
+import org.apache.kafka.streams.processor.StreamPartitioner;
+
+public class RoundRobinPartitioner<K, V> implements StreamPartitioner<K, V> {
+  private final AtomicInteger currentPartition = new AtomicInteger();
+
+  @Override
+  public Integer partition(String topic, K key, V value, int numPartitions) {
+    return currentPartition.getAndSet((currentPartition.get() + 1) % numPartitions);
+  }
+}
